--- Front-End/.gitignore
@@ -1,28 +1,0 @@
-## MAC files ##
-.DS_Store
-
-## NPM ##
-npm
-npm-debug.log
-node_modules/
-
-pids
-*.pid
-logs
-*.log
-
-# Directory for instrumented libs generated by jscoverage/JSCover
-lib-cov
-
-# Coverage directory used by tools like istanbul
-coverage
-
-# Grunt intermediate storage (http://gruntjs.com/creating-plugins#storing-task-files)
-.grunt
-
-# Compiled Dirs (http://nodejs.org/api/addons.html)
-build/
-bundle.js
-
-# aws keys
-config.json
