+++ STYLE-GUIDE.md
@@ -1,0 +1,4 @@
+## Style Guide
+
+Refer to the [AirBnb Style Guide](https://github.com/airbnb/javascript).
+
