--- database/cassandra/cRatings.js
@@ -1,61 +1,0 @@
-const fs = require('fs');
-const path = require('path');
-const createCsvWriter = require('csv-writer').createObjectCsvWriter;
-const faker = require('faker')
-const data = require('./restaurantList.js')
-// restaurantList contains restaurant names from another file
-
-const num_of_records = 100;
-const num_of_docs = 5;
-const doc_size = num_of_records / num_of_docs;
-const restaurantList = data.names
-
-
-const reviewGenerator = async (index) => {
-  const records = [];
-  for (var i = 0; i < doc_size; i++) {
-    records.push({
-      restaurant_id: Math.floor(Math.random() * (2500000 - 1)) + 1,
-      restaurant_name: restaurantList[Math.floor(Math.random() * restaurantList.length)],
-      create_date: faker.date.past,
-      number_of_reviews: Math.ceil(Math.random() * 200),
-      overall_rating: Math.ceil(Math.random() * 50) / 10,
-      one_star_rating: Math.floor(Math.random()*(100)) / 10,
-      two_star_rating: Math.floor(Math.random()*(100)) / 10,
-      three_star_rating: Math.floor(Math.random()*(100)) / 10,
-      four_star_rating: Math.floor(Math.random()*(100)) / 10,
-      five_star_rating: Math.floor(Math.random()*(100)) / 10
-    })
-  }
-
-  const reviewWriter = createCsvWriter({
-    path: path.join(__dirname, 'cCSV', `Creview_${index}.csv`),
-    header: [
-      {id: 'restaurant_id', title:'restaurant_id'},
-      {id: 'restaurant_name', title:'restaurant_name'},
-      {id: 'create_date', title:'create_date'},
-      {id: 'number_of_reviews', title:'number_of_reviews'},
-      {id: 'overall_rating', title:'overall_rating'},
-      {id: 'one_star_rating', title: 'one_star_rating'},
-      {id: 'two_star_rating', title: 'two_star_rating'},
-      {id: 'three_star_rating', title: 'three_star_rating'},
-      {id: 'four_star_rating', title: 'four_star_rating'},
-      {id: 'five_star_rating', title: 'five_star_rating'},
-    ]
-  })
-
-  reviewWriter.writeRecords(records)
-    .then(() => {
-      console.log ('pg review generated done')
-    })
-}
-
-const csvGenerator = async () => {
-  for (var i = 0; i < num_of_docs; i++) {
-    await reviewGenerator(i)
-  }
-}
-
-csvGenerator();
-
-
