--- database/cassandra/restaurantList.js
@@ -1,1526 +1,0 @@
-module.exports.names = [
-  "California Shack",
-  "Chez Panisse",
-  "The French Laundry",
-  "Manresa",
-  "Ember Restaurant",
-  "Mini Kabob",
-  "Zoftig Eatery",
-  "Pink’s Hot Dogs",
-  "Carthay Circle Restaurant",
-  "Tartine Manufactory",
-  "Sam’s Chowder House",
-  "Hog Island Oyster Co.",
-  "Farmhouse Inn Restaurant",
-  "Manhattan Restaurant",
-  "Saltie Girl",
-  "Mamma Maria",
-  "Atlantic Fish Co",
-  "The Capital Grille",
-  "Mario’s Restaurant",
-  "Theo’s Cozy Corner Restaurant",
-  "Rich Table",
-  "Catch 35",
-  "Blue Plate",
-  "California Pizza kitchen",
-  "Desi Addicts",
-  "Mama’s Fish House",
-  "Hot Lunch",
-  "Break Fast",
-  "Spicy Foods",
-  "Tasty Dots",
-  "Taste Corner",
-  "Dinner Side",
-  "Smart Kitchen",
-  "Spicy Pizza Kitchen",
-  "Super Supper",
-  "Sandwich corner",
-  "Divine lunch",
-  "Delighted Kitchen",
-  "Hot & Crusty",
-  "King and Queen",
-  "Blue Collar",
-  "Old Lisbon",
-  "The Marine Room",
-  "Fish Kitchen",
-  "The Smoking Goat",
-  "The Smoking Fish",
-  "Dragon Eats",
-  "Jar + Fork",
-  "Project Juice",
-  "The Pink Door",
-  "Little Sheep",
-  "Big Sheep",
-  "Green Leaf",
-  "Umami Burger",
-  "Burger King",
-  "Golden Burgers",
-  "Fare Start",
-  "The Kitchen",
-  "Indian Food",
-  "Spicy Indian Food",
-  "Rice Corner",
-  "Barley Mash",
-  "Le Bernardin",
-  "Per Se",
-  "Gramercy Tavern",
-  "Barley Mash",
-  "Blunch",
-  "Boka",
-  "Cafe Provence",
-  "Coaster Saloon",
-  "The NoMad Restaurant",
-  "The Modern",
-  "Hollywood Cafe",
-  "Just Falafs",
-  "Lionfish",
-  "Maza Grill Kabob",
-  "Mister A’s",
-  "Palomino",
-  "Buddakan",
-  "THE GRILL",
-  "Via Carota",
-  "Party Fowl",
-  "Peking Inn",
-  "Phat Phuc Noodle Bar",
-  "Pu Pu Hot Pot",
-  "Frenchette",
-  "Jungsik",
-  "Upland",
-  "Blue Hill",
-  "Eleven Madison Park",
-  "Jean-Georges",
-  "Crabby Dick’s",
-  "Fare Start",
-  "Greens",
-  "Gringos Locos",
-  "Hometown Bar-B-Que",
-  "Benoit New York",
-  "Piccola Cucina Estiatorio",
-  "Cloak & Petal",
-  "Corridor",
-  "Duke’s Seafood",
-  "Emerald Grill",
-  "Ocean Prime",
-  "Asiate",
-  "Hearth",
-  "Eatmore Fried Chicken",
-  "Fishcotheque",
-  "Great Eastern Restaurant",
-  "Hillstone",
-  "Island Prime",
-  "Hard Rock Café",
-  "Fiddler’s Green",
-  "Girl & the Goat",
-  "Jar + Fork",
-  "Juan in a Million",
-  "Lox Stock & Bagel",
-  "Del Posto",
-  "Bouley at Home",
-  "Carmine’s Italian Restaurant",
-  "21 Club",
-  "Firefly Restaurant",
-  "Geronimo",
-  "Goldfinch Tavern",
-  "Homestyle Hawaiian",
-  "Kum Den Bar & Restaurant",
-  "The View Restaurant & Lounge",
-  "Trinity Place",
-  "The Polo Bar",
-  "Season 53",
-  "Smyth",
-  "Spinasse",
-  "The French Gourmet",
-  "The Polo Bar",
-  "Oceana Restaurant",
-  "Patsy’s Italian Restaurant",
-  "Staple & Fancy",
-  "Thai Me Up",
-  "Thai The Knot",
-  "The Chef in the Hat",
-  "Ai Fiori",
-  "Junior’s Restaurant & Bakery",
-  "Candle 79",
-  "PM Fish & Steak House",
-  "Project Juice",
-  "Shaker + Spear",
-  "Smoque BBQ",
-  "Ellen’s Stardust Diner",
-  "410",
-  "1919",
-  "SAUTE",
-  "#1 CHINESE KITCHEN",
-  "#1 chinese restaurant",
-  "19TH HOLE",
-  "3 KINGS",
-  "3 MILES HOUSE, INC.",
-  "3 W'S TAVERN",
-  "300 SOUTH ANN STREET",
-  "438 CLUB",
-  "5-MILE HOUSE",
-  "743 S. MONTFORD,INC.",
-  "A & W RESTAURANT",
-  "A TASTE OF CHINA",
-  "ABACROMBIE FINE FOODS",
-  "ABC SUSHI",
-  "ACROPOLIS RESTAURANT",
-  "ADMIRAL FELL INN",
-  "AG & HARP'S",
-  "AIRPORT BAR & GRILL, INC.",
-  "AKBAR RESTAURANT",
-  "AKDEY SUBWAY, INC.",
-  "ALDO'S   RESTAURANT",
-  "ALE MARY'S",
-  "ALEXANDER'S TAVERN",
-  "ALFEO'S",
-  "AL-HO CLUB",
-  "BAY ATLANTIC CLUB",
-  "BAY CAFE",
-  "BEACH  BUMS CAF",
-  "BECKER'S TAVERN",
-  "BEDROCK",
-  "BEDROCK (KITCHEN AREA)",
-  "BELAIR PUB",
-  "BELT STREET PUB",
-  "BELVEDERE TOWERS CAF",
-  "BERGER'S TAVERN, INCORPORATED",
-  "BERMUDA BAR",
-  "BERTHA'S RESTAURANT",
-  "BERT'S EATS & DRINKS (3103-05)",
-  "BEST WESTERN TRAVEL PLAZA",
-  "BIG APPLE TREE CAF",
-  "BIG BAD WOLF'S HOUSE OF BARBEQUE",
-  "BIG KAHUNA CANTINA",
-  "BIG TEE'S BAR & GRILL",
-  "BILL'S 60'S & 70'S CLUB",
-  "BILL'S CAFE",
-  "BILL'S LIGHTHOUSE INN",
-  "BILTMORE SUITES HOTEL",
-  "BIRCHES RESTAURANT",
-  "BIRDS OF A FEATHER",
-  "BLACKTIE SERVICE, LLP",
-  "BLACKTIE SERVICES",
-  "BLARNEY STONE INN",
-  "BLARNEY STONE PUB",
-  "BLEACHERS",
-  "ALLEN RESTAURANT",
-  "ALOHA TOKYO",
-  "ALONSO'S RESTAURANT & COCKTAIL LOUNGE",
-  "AMBASSADOR DINNING ROOM",
-  "AMBROSIA'S",
-  "AMERICAN FOOD SERVICE",
-  "AMERICAN LEGION FEDERAL POST #19",
-  "AMERICAN LEGION POST #187",
-  "AMERICAN LEIGON FORT MCHENRY POST #133",
-  "AMIGO'S BAR",
-  "ANGELINA'S",
-  "ANGELO'S RESTAURANT & CARRYOUT",
-  "ANGELS ROCK BAR",
-  "ANGLE INN",
-  "ANNABEL LEE TAVERN",
-  "ANNS AT LEXINGTON PLACE",
-  "APPLEBEE'S NEIGHBORHOOD GRILL & BAR",
-  "aramark at citi bank baltimore",
-  "ARAMARK FHC SPORTS  (MAIN KITCHEN)",
-  "ARAMARK FHC SPORTS (BAR 200-201)",
-  "ARAMARK FHC SPORTS (BAR 210)",
-  "ARAMARK FHC SPORTS (BAR 217)",
-  "ARAMARK FHC SPORTS (BAR 221)",
-  "ARAMARK FHC SPORTS (BAR 232)",
-  "ARAMARK FHC SPORTS (BAR 236)",
-  "ARAMARK FHC SPORTS (BAR 244)",
-  "ARAMARK FHC SPORTS (BAR 252-253)",
-  "ARAMARK/PNC",
-  "ARA-MONUMENTAL LIFE INSURANCE COMPANY",
-  "ARCH SOCIAL CLUB",
-  "ARCOS",
-  "ARIZONA BAR & GRILL",
-  "ASAHI SUSHI",
-  "asian taste",
-  "ATWATER'S",
-  "AU BON PAIN",
-  "AUSTIN GRILL",
-  "AUSTIN GRILL",
-  "AVENUE BAR & LIQUOR",
-  "azteca de oro",
-  "B",
-  "B & J RESTAURANT",
-  "B & O CAFE",
-  "B G'S",
-  "B.G.'S,PUB INCORPORATED",
-  "babalu grill",
-  "BACKSTRETCH SALOON",
-  "BAJA BEACH CLUB  (COMBINED WITH BAR-L)",
-  "BALTIMORE BEVERAGE CO. LLC",
-  "BALTIMORE BEVERAGE CO. -STARBUCKS",
-  "BALTIMORE COUNTRY CLUB",
-  "BALTIMORE DELIGHT RESTAURANT",
-  "BALTIMORE EAGLE, INC.",
-  "BALTIMORE ESTONIAN SOCIETY",
-  "BALTIMORE MARRIOTT WATERFRONT/EMPLOYEE CAFETERIA",
-  "BALTIMORE MARRIOTT WATERFRONT/GRILLE 700 RESTAURAN",
-  "BALTIMORE MARRIOTT WATERFRONT/KOZMOS BAR",
-  "BALTIMORE MARRIOTT WATERFRONT/RIGANOS DELI",
-  "BALTIMORE MARRIOTT-INNER HARBOR/EMP. CAFETERIA",
-  "BALTIMORE MARRIOTT-INNER HARBOR/REST/BAR",
-  "BALTIMORE PHO",
-  "BALTMORE KICKERS",
-  "BANJARA",
-  "BANTHAI",
-  "BARISTA ESPRESSO CAF",
-  "BARNES & NOBLE BOOKSELLERS",
-  "BARTENDERS",
-  "BLU BAMBU-FRESH AISIAN EATS",
-  "BLUE AGAVE RESTAURANT AND TEQUILERIA",
-  "BLUE CRAB EXPRESS & CAF INC.",
-  "blue island malaysian cuisine",
-  "BLUE JEAN CAFE",
-  "BLUE MOON  DINING  HOUSE  & EXPRESSO  BAR",
-  "BLUE PARROT",
-  "BLUE POINT CRAB HOUSE (seasonal)",
-  "BLUE SEA GRILL",
-  "BLUE SKY RESTAURANT",
-  "BLUEHOUSE",
-  "BO BROOKS",
-  "BOB CUMMINGS PLACE",
-  "BOB DELI",
-  "BOB'S BAR",
-  "BOB'S OLDIE'S CAF",
-  "BOCCACCIO",
-  "BOP BRICK OVEN PIZZA",
-  "BOUILLABAISSE CAF",
-  "BOULDER GARDEN CAFE",
-  "BRASS MONKEY SALOON",
-  "BRASSERIE TATIN",
-  "BREEZE INC.",
-  "BRENDAL, INC.",
-  "BREWER'S HILL PUB & GRILL",
-  "BREWSTER'S TAVERN",
-  "BROADWAY DINER",
-  "BROOKEY'S",
-  "BROOKLYN BOOSTER CLUB",
-  "BROOKLYN INN, LLC",
-  "BROOKLYN PLEASURE CLUB",
-  "BROTHERS PLACE, INC.",
-  "BRUCE MCKELDIN CONSULTING, LLC",
-  "BUD PAOLINO'S RESTAURANT",
-  "BUFFALO WINGGS AND BEER",
-  "BULL ROBINSON  LIQUORS (AKA 2736 GREENMOUNT AV",
-  "BULLSEYE BAR & GRILLE",
-  "BURGER KING",
-  "BURGER KING",
-  "BURGER KING",
-  "BURGER KING",
-  "BURGER KING # 10293",
-  "BURGER KING # 13360",
-  "BURGER KING #10294",
-  "BURGER KING #10295",
-  "BURGER KING #1950",
-  "BURGER KING #2610",
-  "BURKE'S CAFE, INCORPORATED",
-  "BURKE'S CAF/ COMEDY FACTORY",
-  "BUSTER'S PUB",
-  "BUTTS & BETTY'S TAVERN, INC.",
-  "B-WATCH",
-  "BYZANTIO CAF",
-  "CAF AZAFRAN",
-  "CAF EXPRESS",
-  "CAF GIA",
-  "CAF GOURMET",
-  "CAFE' INTERNATIONALE",
-  "CAF LATTE 'DA",
-  "CAFE MANET",
-  "CAF ON THE SQUARE, LLC",
-  "CAF SUSSIE'S",
-  "CAFE ZEN",
-  "CALIFORNIA PIZZA KITCHEN",
-  "CAMDEN PUB",
-  "CAMERON'S SEAFOOD",
-  "CANTINA LAS AMERICAS/LA RUMBA",
-  "CANTON COVE",
-  "CANTON STATION",
-  "CANTON'S PEARLS",
-  "CANTON'S PORTSIDE TAVERN",
-  "CAPTAIN LARRY'S",
-  "caribbean paradise",
-  "CARIBOU COFFEE",
-  "CARL REFEERS BAR & GRILL",
-  "CARLOS O'CHARLIES",
-  "CARL'S LITTLE HOUSE",
-  "CARLYLE CLUB",
-  "CAROLINAS TEX- MEX RESTAURANT",
-  "CAROLINA'S TEX-MEX RESTURANT",
-  "CARROLL STATION",
-  "CARROLL STATION",
-  "CASSANOVIA",
-  "CASTAWAY'S BAR &  GRILL",
-  "CATALINA RESTAURANT",
-  "CATCH 22 SALOON",
-  "CATON HOUSE RESTAURANT",
-  "CATTLE & FIN GRILL",
-  "CAZBAR",
-  "CC LOUNGE",
-  "CEASAR'S DEN",
-  "CEDONIA INN",
-  "CENTER CLUB, INCORPORATED",
-  "CENTER STAGE",
-  "CENTRAL STATION",
-  "CERIELLO FINE FOODS",
-  "CHAMPAGNE TONY'S",
-  "CHARLES & MADISON, LLC",
-  "CHARLES RESTAURANT  &  CARRYOUT",
-  "CHARLES ST DELI",
-  "CHARLES VILLAGE PUB",
-  "CHARLESTON",
-  "CHARLIE & DEE RESTAURANT",
-  "CHARLOTTE'S",
-  "CHARM CITY INN",
-  "CHAZZ &  REN'E",
-  "CHEERLEADERS",
-  "CHEERLEADERS SPORTS BAR",
-  "CHEF PAOLINO",
-  "CHESAPEAKE BAY RESTAURANT",
-  "CHESAPEAKE WINE COMPANY",
-  "CHIAPPARELLI'S RESTAURANT",
-  "CHIC & TROUT (1007-09)",
-  "CHICKEN RICO",
-  "CHIMES CAF",
-  "CHINA CHEF",
-  "CHINA EAST",
-  "CHINA EUTAW EXPRESS",
-  "CHINA GARDEN",
-  "CHINA WOK",
-  "CHINATOWN CAF",
-  "CHINESE DELIGHT",
-  "CHIPOTLE",
-  "CHIPOTLE MEXICAN GRILL",
-  "CHIU'S SUSHI JAPANESE RESTAURANT",
-  "CHIYO SUSHI",
-  "CHOCOLATEA",
-  "CHOPPER'S TAVERN",
-  "CHOW  MEIN CHARLIES  chinese food",
-  "CHRIS LIQUOR & GRILL",
-  "CHUBBIES CLUB",
-  "CHUCK'S",
-  "CHUNG MEE RESTAURANT",
-  "CIAO BELLA RESTAURANT",
-  "CINCO DE MAYO",
-  "CINGHIALE",
-  "CIRCUS BAR & RESTAURANT",
-  "CITY LIMITS SPORTS BAR",
-  "claddagh pub",
-  "CLASSIC CAFE EXPRESS",
-  "CLAYTON FINE BOOKS & CAF",
-  "CLIFTON PARK GOLF COURSE SNACK BAR",
-  "CLIPPER MILL INN",
-  "CLUB 1722",
-  "CLUB BAR",
-  "CLUB BUNNS (606-08)",
-  "club charles zodiac     (1724-1726 n. charles st)",
-  "CLUB CHOICES",
-  "CLUB CLIMAX",
-  "CLUB CONFETTI",
-  "CLUB HAREM",
-  "CLUB HOUSE BAR & GRILL",
-  "CLUB HUSH",
-  "CLUB INTERNATIONAL",
-  "CLUB LUZERENE",
-  "CLUB MATE",
-  "CLUB MIAMI",
-  "CLUB MITCHELL",
-  "CLUB OKAPI",
-  "CLUB ONE",
-  "CLUB PHOENIX",
-  "CLUB PUSSYCAT",
-  "CLUB RIO & TOUCH OF CLASS",
-  "CLUBHOUSE BAR",
-  "CLUBHOUSE BAR & GRILL",
-  "COACH HOUSE INN",
-  "COCKY LOU'S",
-  "COCO MOKA CAF",
-  "COCONUT CAF",
-  "COFFEE LAND",
-  "COLD STONE CREAMERY #1479",
-  "COLD STONE CREAMERY# 102",
-  "COLLINS CORNER",
-  "COLONIAL INN",
-  "COMMON GROUND",
-  "COOKIE'S  CITY  LINE  DINER",
-  "COOPER'S LIQUOR (568-570)",
-  "COPRA",
-  "CORKS",
-  "CORKY'S GRILL",
-  "CORRADO'S CAF",
-  "COSI, INC. # 157",
-  "COSMOPOLITAN BAR & GRILL",
-  "COSMOS RESTAURANT-CARRYOUT (5427-29)",
-  "COTTON CLUB",
-  "courthouse east caf‚'",
-  "COURTYARD BY MARRIOTT & COURTYARD CAFE & LOUNGE",
-  "crazy johns",
-  "CRAZY LIL'S",
-  "CREATIVE ALLIANCE",
-  "CREPE DU JOUR",
-  "CROSSROADS SPORTS PUB & GRILL",
-  "CURB SHOPPE",
-  "CURLEY ST.  CAF",
-  "CURTIS BAY ATHLETIC CLUB",
-  "D J'S GASLIGHT",
-  "DAILY  GRIND  COFFEE  HOUSE",
-  "DAILY GRIND",
-  "DAILY GRIND",
-  "dalesio's of little italy llc",
-  "DANCING PIGS",
-  "DANNY'S LOUNGE",
-  "DARBY'S",
-  "DARKER THAN BLUE",
-  "DASH SEAFOOD & CHICKEN",
-  "DAVEY JONES LOCKER",
-  "DAVID & DAD'S LLC",
-  "DAVID'S RESTAURANT & DELI",
-  "DAYS INNS MANAGEMENT CO",
-  "DEAD END SALOON",
-  "DEAD FREDDIES",
-  "DECADES SPORTS BAR",
-  "DECKERS BAR & GRILLE",
-  "DEEP SIDE RESTAURANT & CAF",
-  "DEGO DAME'S",
-  "DELLA ROSE'S 2 CANTON CROSSING",
-  "DESERT CAFE'(1605-07)",
-  "DESERT INN",
-  "DIAMOND LOUNGE   (415 1/2 E. BALTIMORE STREET)",
-  "DIANNE'S RESTAURANT LOUNGE",
-  "DILONE GROCERY & RESTAURANT",
-  "DIMIS CAF AND GRILL, INC.",
-  "DIMITRI'S TAVERN",
-  "DING HOW RESTAURANT",
-  "DIONYSUS RESTAURANT & LOUNGE",
-  "DIZZY  IZZY",
-  "DOMINICANO INTERNACIONAL",
-  "DOMINO SUGAR CAFATERIA OPERATION # 65227",
-  "DONNA'S",
-  "DONNA'S",
-  "DONNA'S",
-  "DONNA'S CHARLES VILLAGE, LLC",
-  "DONNA'S CROSS KEYS",
-  "DONRITZ BAR & GRILL",
-  "DOUGHERTY'S",
-  "DOWN THE HATCH",
-  "DOWNTOWN SPORTS EXCHANGE",
-  "DREAM TAVERN, CARRIBEAN  U.S.A.",
-  "DREAMS INN",
-  "DRIFTERS RAW BAR & GRILL",
-  "DUCK'S BLACK SATIN LOUNGE",
-  "DUCKY PLEASURE CLUB",
-  "DUCLAW BREWING COMPANY",
-  "DUDA'S TAVERN",
-  "DUKEM ETHIOPIAN RESTAURANT",
-  "DUNKIN DONUTS",
-  "Dunkin Donuts",
-  "DUNKIN DONUTS",
-  "DUNKIN DONUTS",
-  "DUNKIN DONUTS",
-  "DUNKIN DONUTS & BASKIN ROBBINS",
-  "DUNKIN DONUTS/BASKIN ROBBINS",
-  "DUTCH POT CAF",
-  "DYNASTY",
-  "E C CAF & CATERING",
-  "EAGLE NEST",
-  "EAST BALTIMORE ENTERPRISE",
-  "EASTERN HOUSE RESTAURANT",
-  "EASTSIDE  SPORTS  SOCIAL  CLUB",
-  "EAT MUST BE FIRST",
-  "ECLECTIC CAF",
-  "edelweiss bakery & caf‚'",
-  "EDENS LOUNGE",
-  "EDGAR'S INC.--BALTIMORE, MD",
-  "EDIE'S DELI & GRILL",
-  "EDISON   LOUNGE",
-  "EDO SUSHI",
-  "EICHENKRANZ RESTAURANT",
-  "EL OSASIS RESTAURANT & DISCOTEC",
-  "EL RANCHO BLANCO",
-  "EL RANCHO GRANDE",
-  "EL RECODO, INC.",
-  "EL RINCONCITO PERUANO",
-  "EL TORITO MEXICAN RESTAURANT",
-  "EL TROPICAL RESTAURANT (3736-38)",
-  "ELDORADO INC.",
-  "ELKS I.B.P.O.E.OF W.",
-  "ELLIOT'S POUR HOUSE",
-  "EL-MOLINO",
-  "ELVIS RESTAURANT",
-  "ENRICO'S SPORTS CAFE",
-  "ESPN  ZONE",
-  "europe restaurant & catering",
-  "EVERYDAY'S LLC",
-  "EXCLUSIVE",
-  "family diner",
-  "FAMOUS RESTAURANT",
-  "FANTASIES",
-  "FANTASY CAF",
-  "FEDERAL HILL LOUNGE",
-  "FELL'S POINT CARRYOUT",
-  "FIELDER'S TAVERN",
-  "FIELDS OLD TRAIL",
-  "FIELDS RIB HOUSE",
-  "FILLIN' STATION",
-  "FIRESIDE NORTH LIQUORS",
-  "FISHERMANS WHARF INNER HARBOR",
-  "FIVE GUYS @ BREWERS HILL",
-  "FIVE GUYS FAMOUS BURGERS & FRIES",
-  "FIVE GUYS FAMOUS BURGERS & FRIES",
-  "FIVE POINT TAVERN",
-  "FIVE SEASONS",
-  "FLAMERS BURGER & CHICKEN",
-  "FLAMINGO  LOUNGE",
-  "Fleming's Prime Steakhouse and Wine Bar",
-  "FLETCHERS",
-  "FLIP'S AT CLIPPER MILL",
-  "FOGO DE CHAO",
-  "FOODTOPIA",
-  "FORT CHARLES PUB",
-  "FOUL BALL BAR",
-  "FOUR G'S LIQUOR & LOUNGE",
-  "FOXY LADY",
-  "FRANKLINTOWN INN",
-  "FRANK'S PIZZA",
-  "fraternal order of police baltimore city lodge #3",
-  "FRAZIER'S ON THE AVENUE",
-  "FRED & MARGIE'S INN",
-  "FREDERICK STATION SALOON",
-  "FRIENDLY  FRIED  CHICKEN",
-  "FRIEND'S RESTAURANT AND SPIRITS",
-  "FULL MOON SALOON",
-  "FULTON RESTAURANT",
-  "FULTON RESTAURANT    (2201 1/2 FULTON AVENUE)",
-  "G & A RESTAURANT",
-  "GAFFNEY'S BACK FIN ( seasonal)",
-  "GALLAGHERS BAR & GRILL    (938-940)",
-  "GALLERY",
-  "GALLERY CAF",
-  "GARDEN BEEF & BEER",
-  "GARDEN RESTAURANT",
-  "GECKOS",
-  "GENJI SUSHI EXPRESS   (A.K.A. 600 S EXETER ST)",
-  "GEN-JOSEPH HALLER POST #95, AMERICAN LEGION, INC.",
-  "GENTLEMEN OF CHARISMA",
-  "GEORGE'S CAFETERIA",
-  "GEORGIO'S",
-  "GERMANIO'S RESTAURANT",
-  "GIBSON'S",
-  "GILMOR PLEASURE CLUB",
-  "GINGER'S",
-  "GLADIATORS ASSOCIATIONS, INC",
-  "GLASS DOOR lounge",
-  "GOLDEN KEY RESTAURANT",
-  "GOLDEN NUGGET",
-  "GOLDEN NUGGET LOUNGE",
-  "GOLDEN WEST CAF, INC.",
-  "GOOD TIMES BAR AND LOUNGE",
-  "GORDON'S",
-  "GOSSIP BISTRO",
-  "GRAND CRU",
-  "GRANDVIEW RESTAURANT",
-  "GRANITE",
-  "GRANT'S 2 SPOT",
-  "GREAT TASTE CHINESE  RESTAURANT",
-  "GREENE ST CAF",
-  "GREENHOUSE CAF",
-  "GREENMOUNT LOUNGE",
-  "GRENADIER CLUB",
-  "GRIFFITH'S",
-  "GRILL ART",
-  "GRILLE 601",
-  "GROWLERS PUB",
-  "GROWN FOLKS BAR",
-  "GRUMPIE'S",
-  "GRUNDY'S",
-  "GRUNWALD CLUB, INCORPORATED",
-  "JAUNG KAK RESTURANT(aka 18 W.20TH ST)",
-  "JAVA JOE'S",
-  "JAY & KAY SUBWAY, INC.",
-  "JAY'S ON READ",
-  "JAY'S WOLFE ST CAF",
-  "JAZZY BLUES CLUB",
-  "JD's SMOKEHOUSE BAR & GRILL, INC.",
-  "JEMELLA 'S CUT RATE",
-  "JERRY'S BAR",
-  "JEWEL BOX LOUNGE",
-  "JHJ SALOON",
-  "JIGGS TAVERN",
-  "JIM DANDY'S TAVERN",
-  "JIMMY'S RESTAURANT",
-  "JIM'S DINER",
-  "JOCKEY CLUB/PALACE KITCHEN",
-  "JOE SQUARED PIZZA AND BAR",
-  "JOE'S TAVERN",
-  "JOE'S TAVERN",
-  "JOE'S TAVERN & PACKAGE GOODS",
-  "JOHN'S LOUNGE",
-  "JOY DELI",
-  "JP'S",
-  "JR.'S BAR & GRILL",
-  "JUNIORS",
-  "K.F.C.",
-  "KABOBBI RESTAURANT",
-  "KALI'S COURT",
-  "POSITIONS UNLIMITED",
-  "POTBELLY SANDWICH WORK'S",
-  "PRIMO'S",
-  "PULASKI FOOD SERVICE",
-  "PUR LOUNGE",
-  "PURPLE GOOSE",
-  "q s r,inc  - Mcdonald's",
-  "qsr, inc.- McDonalds",
-  "QSR, INC. MCDONALD'S)",
-  "QUIGLEY 1/2 IRISH PUB",
-  "QUINOS",
-  "QUIZINOS OF FALLS road",
-  "QUIZNOS",
-  "QUIZNO'S",
-  "QUIZNO'S",
-  "QUIZNOS OF FELLS POINT",
-  "QUO VADIS DEMOCRATIC CLUB",
-  "RA SUSHI",
-  "RADISSON CROSS KEYS HOTEL",
-  "RADISSON HOTEL AT CROSS KEYS",
-  "RAFTERS",
-  "RAINBOW HOMESTYLE BUFFET",
-  "RAINBOW'S  BAR",
-  "RALLO'S RESTAURANT",
-  "RAMS HEAD TAVERN",
-  "RANDY'S BAR",
-  "RAS-A-TER INTERNATIONAL",
-  "RASCALS COMEDY CLUB",
-  "RAVE  INN",
-  "RAY LEWIS' FULL MOON B-B-Q",
-  "RAY'S TAVERN",
-  "RAYZER'S",
-  "RED DOOR LOUNGE",
-  "RED FISH",
-  "RED FOX LIQUORS",
-  "RED HOUSE TAVERN INC.",
-  "RED LYON TAVERN",
-  "RED MAPLE",
-  "RED ROOSTER",
-  "RED SQUARE RESTAURANT",
-  "REGAN'S TAVERN",
-  "REGI'S BAR & RESTAURANT",
-  "GUS RESTAURANT (3103-05)",
-  "GYPSIES",
-  "H.I.S. PLACE",
-  "HABENERO GRILL",
-  "HAMILTON HOMING PIGEON CLUB, INC.",
-  "HAMPDEN REPUBLICAN CLUB",
-  "harbor way inn",
-  "HARBORSIDE CHICKEN & SEAFOOD",
-  "HARD ROCK CAF",
-  "HARD ROCK CAF-OUTDOOR PIER",
-  "HARFORD HOUSE",
-  "HARFORD ROAD GROUP, LLC",
-  "HARFORD ROAD LIQUORS",
-  "HARRAR CAF",
-  "MIDWAY BAR",
-  "MIKE MCGOVERN'S",
-  "MIKE'S PLACE",
-  "MILTON AT CHARLES ST",
-  "MINATO RESTAURANT",
-  "MINATO/CAFE VIET RESTAURANT",
-  "MISS SHIRLEY'S CAF",
-  "MOM'S PLACE",
-  "MONDO BONDO ITALIAN BISTRO",
-  "MONUMENT HOUSE",
-  "MORGAN STATE-FACULTY",
-  "MORTON'S--THE  STEAKHOUSE",
-  "MO'S CRAB & PASTA FACTORY",
-  "MOTHER'S  FEDERAL  HILL  GRILL",
-  "MOUNT VERNON CLUB OF BALTIMORE, INC.",
-  "MOUNT VERNON STABLE & SALOON",
-  "MOUSE TRAP",
-  "MR. C. PLACE",
-  "MR. JOE'S BAR AND CUT RATE LIQUORS",
-  "MT VERNON ENTERTAINMENT LLC",
-  "MT. PLEASANT INN",
-  "MT. ROYAL TAVERN",
-  "MT. VERNON  HOTEL AND WASHINGTON CAFE'",
-  "MT. WASHINGTON TAVERN",
-  "MUGGSY'S",
-  "MUGHAL GARDEN",
-  "MUIR'S TAVERN",
-  "MUM'S",
-  "MURPHY'S NEIGHBORHOOD BAR & GRILL",
-  "MUSTANG ALLEY'S",
-  "MUSTANG INN",
-  "MVP ENTERTAINMENT LLC CLUB 347-KITCHEN AREA",
-  "MY COUSIN'S PLACE, LLC (3925-27)",
-  "MY PANINI",
-  "MY THAI RESTAURANT",
-  "NAK WON RESTAURANT",
-  "NAM KANG KOREAN RESTAURANT   (AKA 101 W. 22ND ST.)",
-  "NASU BLANCA",
-  "NCDEVINS",
-  "NELSON BAKED GOODS & COFFEE SHOP",
-  "NEOPOL",
-  "network deli",
-  "NEVER ON SUNDAY",
-  "NEVIN'S  CROSS  STREET STATION",
-  "NEW CLUB THUNDERBIRD INC.",
-  "NEW DUFFY' S RESTAURANT",
-  "NEW EVERGREEN CAF",
-  "NEW HONG  KONG RESTAURANT, INC.",
-  "NEW JOHNSON'S LOUNGE",
-  "NEW MAY FLOWER BUFFET, INC.",
-  "NEW MAYFIELD, INC.",
-  "NEW MOM & DAD LIQUORS",
-  "NEW NO  DO  JI  RESTAURANT",
-  "HARRIS' TAVERN",
-  "HARRY OLIVER MANAGEMENT CO., INC.",
-  "HARRY'S AMERICAN BAR",
-  "HARVEST TABLE",
-  "HAVE A NICE DAY CAF  (AKA 2 MARKET PLACE)",
-  "HAVEN PLACE",
-  "hawley's  (2616-20)",
-  "HEATH STREET PUB",
-  "HEIGHTS LIQUORS, INC.",
-  "HELEN'S GARDEN  CAFE'",
-  "HELLAS COFFEE SHOP",
-  "HENNINGER'S TAVERN",
-  "HIDING SPOT, INC.",
-  "HIGH GROUNDS COFFEE ROASTERS, LLC",
-  "HIGHLAND CLIPPER CLUB",
-  "HILLEN RESTAURANT",
-  "HIPPOPOTAMUS (932-34)",
-  "HOGAN'S ALLEY",
-  "HOLIDAY HOUSE",
-  "HOLIDAY INN-INNER HARBOR",
-  "HOLIDAY RESTAURANT",
-  "HOLLINS STREET PUB",
-  "HOLLINSWOOD INN",
-  "HOLLYWOOD DINER",
-  "HOLY FRIJOLES",
-  "HOMETOWN DELI",
-  "HOMETOWN GIRL & COMPANY",
-  "HONEY'S LOUNGE",
-  "HOOTERS OF AMERICA",
-  "HOPKINS DELI",
-  "HOPKINS PLAZA DELI",
-  "HOPKINS SUBWAY",
-  "HORSEFEATHER'S PUB",
-  "HOT TOMATOES-CAF FORTUNATO",
-  "HOULIHANS OF MARYLAND",
-  "HOWL AT THE MOON  (AKA 22 MARKET PL)",
-  "HUCKAS",
-  "HULL STREET BLUES",
-  "HUNAN EMPIRE",
-  "HYATT REGENCY BALTIMORE      (PISCES & BAR)",
-  "HYATT REGENCY BALTIMORE    (Perks -LOBBY BAR)",
-  "HYATT REGENCY BALTIMORE   (CAFETERIA)",
-  "HYATT REGENCY BALTIMORE  (KITCHEN/BANQUET ROOM)",
-  "HYATT REGENCY OF BALTIMORE  (Bistro-300 & Bar)",
-  "I BAR",
-  "I.B.P.O.E. OF W., INCORPORATED",
-  "I'CON INN",
-  "IGGIES",
-  "IGUANA CANTINA BAR",
-  "IKAN SEAFOOD",
-  "IKAROS RESTAURANT",
-  "IKE'S PLACE",
-  "IL MONDO,INC",
-  "ILLUSIONS MAGIC BAR & LOUNGE",
-  "IMPRESSED CAF",
-  "INCOGNITO TAVERN",
-  "INDIA RASOI RESTAURANT",
-  "INDIA TANDOOR",
-  "INDIGMA",
-  "INSIDE RAIL BAR",
-  "INTERCONTINENTAL HARBOR COURT-BRIGHTON'S",
-  "INTERCONTINENTAL HARBOR-EXPRESSO",
-  "INTRINSICS RESTAURANT",
-  "IRISH PUB",
-  "ISIS ( 226-228)",
-  "ISLAND II",
-  "ITALIANOS RESTAURANT CARRYOUT",
-  "ixia restaurant",
-  "J & J LOUNGE",
-  "J & J TAVERN",
-  "J & M CORNER BAR",
-  "J & S BAR & LOUNGE",
-  "J RAY'S CAFE",
-  "J. PATRICK, INC.",
-  "J. PAUL'S",
-  "J. T.'S SALOON",
-  "JACK'S BISTRO",
-  "JACK'S LIQUOR & LOUNGE",
-  "JAMES JOYCE IRISH PUB AND RESTAURANT",
-  "KAMP  (2312-14)",
-  "KASBAH CAF",
-  "KAWASAKI CAF",
-  "KAY-CEE CLUB",
-  "K'C' CORNER TAVERN",
-  "KELLY'S",
-  "KELLY'S BAR & LOUNGE",
-  "KENTUCKY  FRIED  CHICKEN",
-  "KENTUCKY FRIED CHICKEN",
-  "KENTUCKY FRIED CHICKEN",
-  "KENTUCKY FRIED CHICKEN",
-  "KENTUCKY FRIED CHICKEN",
-  "KENTUCKY FRIED CHICKEN",
-  "KENTUCKY FRIED CHICKEN  Y070-074",
-  "KFC / LONG JOHN SILVERS",
-  "KFC-Y320011",
-  "KICK BACK JACK'S",
-  "KIKU SUSHI RESTAURANT",
-  "KIMMY'S RESTAURANT & CARRYOUT, INC.",
-  "KIM'S LIQUORS",
-  "KIM'S RESTAURANT & CARRYOUT",
-  "KING BUFFET",
-  "KING'S GARDEN CAF, LLC",
-  "KING'S PIZZA EXPRESS",
-  "KIRBIE'S GRILL & CAF",
-  "KIRBY'S KORNER",
-  "KISLING'S TAVERN",
-  "KISS CAF",
-  "KITTY'S LOUNGE",
-  "K-MART MORAVIA  STORE #3229",
-  "KNOTTY PINE INN",
-  "KOBA CAF",
-  "KOCO'S PUB, INC.",
-  "KOLPERS RESTAURANT AND BAR",
-  "KOOPER'S TAVERN",
-  "KOSHER BITE",
-  "KOZY KORNER",
-  "KRISTO'S CARRYOUT",
-  "K'S KORNER",
-  "KUANG NING CARRYOUT",
-  "KUMARI RESTAURANT & BAR",
-  "Kwang Moo Liquor",
-  "L. P. DOCKS",
-  "L.P. STEAMERS",
-  "LA  CAZUELA RESTAURANT",
-  "LA MARSA",
-  "LA ROCA",
-  "LA ROCA BRUJA",
-  "LA RUMBA",
-  "LA SCALA RISTORANTE ITALIANO",
-  "LA SIRENITA RESTAURANT, II",
-  "LA TAVOLA RESTAURANT",
-  "LADY D'S CAF",
-  "LAGERS PUB",
-  "LAHOR KABAB HOUSE",
-  "LA'RAE",
-  "LARRY FLYNT'S HUSTLER CLUB",
-  "LAS ESPERANZA",
-  "LAS PALMAS MEXICAN RESTAURANT, INC.",
-  "LATASCA",
-  "LATIN PALACE  (509-13)",
-  "LATIN QUARTERS",
-  "LAVA LOUNGE",
-  "LEADBETTERS II",
-  "LEAR'S II",
-  "LEBANESE TAVERNA   (AKA 719 S. PRESIDENT STREET)",
-  "LED ENTERPRISES LTD.",
-  "LEGAL SEA FOODS",
-  "LEGEND RESTAURANT",
-  "LEMONGRASS BALTIMORE",
-  "LENNY'S DELI OF LOMBARD STREET",
-  "LEN'S SANDWICH SHOP",
-  "LEON'S BAR",
-  "LEVEL'S LOUNGE",
-  "LIGHT STREET CAF - 17th FLOOR",
-  "LIGHT STREET TAVERN",
-  "LIKE FATHER AND SON",
-  "LIL'S PLACE",
-  "LIME",
-  "LINDEN DELI",
-  "LINDEN LOUNGE & BAR",
-  "LITHUANIAN HALL ASSOCIATION, INC.",
-  "LITTLE BEAVER SOCIAL CLUB",
-  "LITTLE HAVANA",
-  "LITTLE TAVERN SHOPS",
-  "LITTLE TAVERN SHOPS(6412-14)",
-  "LOCO HOMBRE",
-  "LOCUST POINT TAVERN",
-  "LONG HING CARRYOUT",
-  "LONG JOHN SILVERS/A &W RESTAURANTS",
-  "LONG JOHN'S PUB",
-  "LONGWOOD CAFE'",
-  "looney's pub",
-  "LOS MARIACHIS RESTAURANT",
-  "LOUISIANA",
-  "LOUNGE 51",
-  "LOU'S BAR",
-  "LOWELL LIQUOR",
-  "LUCA'S CAF",
-  "LULU'S OFF BROADWAY",
-  "LUMBINI",
-  "LUNA DELSEA",
-  "LUST",
-  "LYRIC THEATRE",
-  "M & J'S SOUL FOOD",
-  "M & M LIQUOR AND LOUNGE",
-  "M & S GRILL/MCCORMICK & SCHMICK'S SEAFOOD REST.",
-  "MACEO'S",
-  "MAGERK'S PUB",
-  "MAGGIE MOORE'S IRISH PUB & RESTAURANT",
-  "mahaffey's pub",
-  "MAKUTI ENTERPRISES, LLC",
-  "MAMA'S ON THE HALF SHELL",
-  "MAMMA LUCIA PIZZA",
-  "MAMMA ROMA",
-  "MAMMA'S CUCINA",
-  "MARGARETS PLACE",
-  "MARIA D'S",
-  "MARTICK'S",
-  "MARYLAND CASUALTY  COMPANY (CAF)",
-  "MARYLAND CLUB",
-  "MARYLAND SCIENCE CENTER BEAKER'S CAF",
-  "MARY'S TAVERN",
-  "MASTELLONE'S",
-  "MATSURI RESTAURANT",
-  "matthew's pizza",
-  "MAX'S ON BROADWAY",
-  "MAYNARD'S CAFE & RESTAURANT",
-  "MBEPB VF #109-PERRY'S CRESENT CAF",
-  "MBEPB VF #137--- STATE CENTER CAF",
-  "MBEPB VF #34",
-  "MCCABE'S",
-  "MCCORMICK & SCHMICK SEAFOOD RESTAURANT",
-  "MCCRACKENS ON KEY",
-  "MCDONALD'S",
-  "mcdonald's",
-  "mcdonald's",
-  "MCDONALD'S",
-  "MCDONALD'S",
-  "MCDONALD'S",
-  "MCDONALD'S",
-  "mcdonald's",
-  "mcdonald's",
-  "MCDONALD'S",
-  "MCDONALD'S",
-  "MCDONALD'S",
-  "MCDONALD'S  RESTAURANT",
-  "MCDONALD'S #5437",
-  "MCDONALD'S (4526-38)",
-  "MCDONALD'S 11851",
-  "MCDONALD'S HAMBURGERS",
-  "MCDONALD'S HAMBURGERS",
-  "MCDONALD'S HAMBURGERS",
-  "MCDONALD'S RESTAURANT",
-  "MCDONALD'S RESTAURANT",
-  "MCHENRY STREET PUB",
-  "MEADOW MILL ATHLETIC CLUB",
-  "MEHEK",
-  "MELI PATISSERIE & BISTRO",
-  "METROPOLITAN COFFEEHOUSE AND WINE BAR",
-  "MEX RESTAURANT",
-  "MICHAEL'S STEAK & LOBSTER HOUSE",
-  "MICHELLE'S CAF",
-  "MICK 0' SHEA'S IRISH PUB",
-  "MIDTOWN YACHT CLUB",
-  "NEW REX LIQUORS,INC.",
-  "NEW RITZ CABARET",
-  "NEW SEOUL RESTAURANT",
-  "NICHIBAN OF FEDERAL HILL, INC.",
-  "NICK'S",
-  "NICK'S FISH HOUSE",
-  "NIGHT OF THE COOKERS. LLC",
-  "NIGHTSHIFT",
-  "NIWANA",
-  "NO IDEA",
-  "NO WAY JOSE CAF",
-  "NO. #1 CHINESE RESTAURANT",
-  "NORMA  JEANS",
-  "NORTH POLE ICE CREAM COMPANY",
-  "NORTH'S CAF",
-  "NORTHSIDE SPORTS BAR",
-  "OASIS NITECLUB",
-  "O'BRADY SALOON",
-  "O'CONNORS",
-  "O'DONNELL PUB",
-  "ODYSSEY LOUNGE",
-  "O'HARA'S IRISH PUB",
-  "OK TAVERN",
-  "OLD BAY CAF",
-  "OLDE OBRYCKI'S CRAB HOUSE (saesonal)",
-  "OLORI B'",
-  "OMEGA PSI PHI FRATERNITY",
-  "ONE + ONE CARRYOUT",
-  "ONE WORLD CAF",
-  "ONE-EYED MIKE'S",
-  "OPEN HEARTH INN",
-  "ORIENT EXPRESS",
-  "ORLEANS LIQUORS",
-  "ottobar",
-  "OUT TO LUNCH",
-  "OUTBACK STEAKHOUSE",
-  "OVERLEA   DINER",
-  "OXFORD TAVERN",
-  "P & S JAMAICAN & AMERICAN RESTAURNAT & CARRYOUT",
-  "P.J.'S PUB",
-  "PAIK'S PLACE",
-  "PAK'S LIQUORS",
-  "PALOMA'S (351-55)",
-  "PANADERIA HERMANOS NAVARRO LLC",
-  "PANDA EXPRESS # 1152",
-  "PANERA BREAD",
-  "PANERA BREAD",
-  "PAPA  JOHNS  PIZZA",
-  "PAPA JOHN'S",
-  "PAPA JOHN'S PIZZA 730",
-  "PAPERMOON DINER",
-  "PARADISE OF PARK HEIGHTS",
-  "PARADOX",
-  "PARK  AVE.  SPORTS  BAR",
-  "PARK AVENUE CAF EXPRESS",
-  "PARK INN",
-  "Parkside Restaurant",
-  "PARTHENON DINER",
-  "PASTICCIO ITALIAN KITCHEN",
-  "PATRICK'S  OF PRATT STREET",
-  "PATTERSON PERK, LLC",
-  "PATTY CAKES",
-  "PAUL'S",
-  "PAZO",
-  "PAZZA LUNA",
-  "peabody court a clairon hotel",
-  "PEARL'S",
-  "PEIWEI ASIAN DINER",
-  "peking house",
-  "PENN LIQUORS",
-  "PENN RESTAURANT & CARRYOUT",
-  "PENN STREET TAVERN",
-  "PENNINGTON   STATION",
-  "PENNINGTON'S PUB      (5 & 7 W. PATAPSCO AVE.)",
-  "PEPINO'S INCORPORATED",
-  "PEPPER'S LIQUOR'S",
-  "PERGUSA ENTERPRISES",
-  "PETALO'S",
-  "PETERS POUR HOUSE AND PUB 110",
-  "PETE'S GRILLE",
-  "PETE'S INN, INCORPORATED",
-  "PETIT LOUIS BISTRO",
-  "PETTIFORD'S SEAFOOD & SOUL FOOD",
-  "PF CHANGS CHINA BISTRO",
-  "PHANTOM'S BAR AND GRILL",
-  "PHILLIP'S EXPRESS BUFFET",
-  "PHILLIP'S OF HARBOR PLACE",
-  "PICCADELIS",
-  "PICKLES PUB",
-  "PIER 5 HOTEL",
-  "PIMLICO TRACK KITCHEN  (PIMLICO RACE COURSE)",
-  "PIZZA PRESTO",
-  "PIZZERIA UNO",
-  "PLAZA SALOON, INC.",
-  "POLISH NATIONAL ALLIANCE, COUNCIL 21",
-  "POLLO REAL",
-  "POLOCK JOHNNY'S",
-  "POL'S CAF",
-  "PONCABIRD PUB/SNACK SHACK",
-  "POPEYES FAMOUS FRIED CHICKEN",
-  "POPEYES FAMOUS FRIED CHICKEN",
-  "POPEYES FAMOUS FRIED CHICKEN",
-  "POPEYES FAMOUS FRIED CHICKEN",
-  "POPEYES FAMOUS FRIED CHICKEN",
-  "POPEYES FAMOUS FRIED CHICKEN",
-  "POPEYES FAMOUS FRIED CHICKEN",
-  "POPEYE'S FAMOUS FRIED CHICKEN & BISCUITS",
-  "POPEYES OF MONDAWMIN",
-  "PORT IN A STORM",
-  "PORT TRUCK STOP BAR  RESTAURANT",
-  "PORTERS PUB",
-  "REISTERSTOWN ROAD SUBWAY",
-  "RENAISSANCE HARBORPLACE HOTEL",
-  "RENAISSANCE HARBORPLACE HOTEL",
-  "RENAISSANCE HARBORPLACE HOTEL",
-  "RENAISSANCE HARBORPLACE HOTEL",
-  "RENDEZVOUS  LOUNGE",
-  "restaurante el trovador",
-  "restaurate el salvador",
-  "RICK'S BAR & GRILL",
-  "RITA'S ITALIAN ICE",
-  "ROBBIE'S NEST",
-  "ROBERT OLIVER SEAFOOD",
-  "ROCCO'S CAPRICCIO",
-  "ROCK \"N\" ROBIN CAFE",
-  "ROCKET TO VENUS (3358-3360)",
-  "RODOS  BAR",
-  "ROMAN'S PLACE",
-  "ROOTS UNLIMITED",
-  "ROPEWALK TAVERN",
-  "ROSIE'S",
-  "ROSITAS RESTAURANT",
-  "ROYAL RESTAURANT",
-  "ROYAL VILLAGE",
-  "ROY'S EAST ATLANTIC I, LIMITED PARTNERSHIP",
-  "RUB BARBEQUE",
-  "RUBY TUESDAY",
-  "RUBY TUESDAY",
-  "RUMOR BAR AND GRILL",
-  "RUSSELL SERNOCKY AMERICAN LEGION POST #184",
-  "RUSTY SCUPPER",
-  "RUTH E'S PLACE",
-  "RUTH'S CHRIS STEAK HOUSE",
-  "RUTH'S CHRIS STEAK HOUSE",
-  "RUTLAND BAR",
-  "RYAN'S DAUGHTER",
-  "RYLEIGH'S OYSTERS",
-  "S.O.S BAR & LOUNGE",
-  "SABATINO'S RESTAURANT",
-  "saigon remembered restaurant",
-  "SAINT BENEDICT BOOSTER CLUB",
-  "SALSA RICO",
-  "SALSARITA'S FRESH CANTINA",
-  "SALT",
-  "SAMMY'S TRATTORIA",
-  "SAMOS RESTAURANT",
-  "SAMPSON'S RESTAURANT",
-  "SAM'S BAGELS",
-  "SAND MAN",
-  "SANDPIPER",
-  "SANDY'S CAFE",
-  "SANTA ANA RESTURANT",
-  "SBARRO INC.",
-  "SBARRO ITALIAN EATERY",
-  "SCHAEFER'S BAR & RESTAURANT",
-  "SCORES",
-  "serdenes restaurant",
-  "S'GHETTI EDDIE'S",
-  "SHAFFERS",
-  "SHAHRAZAD'S INTERNATIONAL NEO-SOUL",
-  "SHANE'S SANDWICH SHOP",
-  "SHARKY'S BAR & GRILL",
-  "SHED ROW",
-  "SHELDON'S LOUNGE",
-  "SHERATON BALTIMORE CITY CTR HOTEL (BANQUET)",
-  "SHERATON INNER HARBOR HOTEL - ORIOLES GRILL",
-  "SHERATON INNER HARBOR HOTEL -EMPLOYEE'S CAFETERIA",
-  "SHERATON INNER HARBOR HOTEL-MAIN KITCHEN",
-  "SHERRIES",
-  "SHERRY'S PLACE",
-  "SHIRLEY'S  HONEY HOLE",
-  "SHORTY'S",
-  "SHUCKERS OF FELLS POINT",
-  "SHULA'S @ SHERATON BALTIMORE CITY CTR HOTEL",
-  "SIDE  BAR  (218-220)",
-  "SID'S TAVERN",
-  "SIGN OF THE TIMES  (137-39)",
-  "SIG'S BAR",
-  "SILK ROAD EXPRESS-MATTIN CENTER  J H U",
-  "SILK'S TAVERN",
-  "SILVER STAR RESTAURANT/CARRYOUT",
-  "SIP & BITE RESTAURANT",
-  "SIR'S DELI",
-  "SISSON STREET SUBWAY",
-  "SISTA'  SANDRA'S INTERNATIONAL CUSINE",
-  "SKEETER'S",
-  "SKY LOUNGE/ TANGO TAPAS",
-  "SKYBOX BAR & GRILL",
-  "SLAINTE IRISH PUB & RESTAURANT",
-  "SLIDERS BAR & GRILL",
-  "SLIM ACE OF CLUB",
-  "SLY FOX PUB",
-  "SMEDLY'S",
-  "smiley's tavern",
-  "SMITH PLACE",
-  "SMITTIE'S BAR & LOUNGE",
-  "SOBO CAFE'  (6-8 W. CROSS ST.)",
-  "SODEXHO  sunpapers",
-  "SODEXHO MANAGEMENT, INC.",
-  "SODEXHO MANAGEMENT. INC",
-  "SODEXHO/ FEDERAL RESERVE BANK",
-  "SOFI'S CREPES",
-  "SONAR",
-  "SON-SON LIQUOR & TAVERN",
-  "SOTTO SOPRA",
-  "SOUL SOURCE",
-  "SOUP'S ON BALTIMORE",
-  "SOUTHERN ACCENTS",
-  "SOUTHSIDE SALOON",
-  "SOUVLAKI  PLACE",
-  "SPARTAN PIZZERIA RESTAURANT",
-  "SPEAKEASY SALOON",
-  "SPECTRUM CATERING",
-  "SPIRITS & MORE",
-  "SPIRITS TAVERN",
-  "SPIRITS WEST,INC",
-  "SPOONS COFFEE",
-  "SPORTS CAFE",
-  "SPORTS PALACE",
-  "SPORTSMAN LOUNGE BAR",
-  "SPORTSMAN SPOT",
-  "SPORTSTERS, INC.",
-  "ST. MARY'S (118-120)",
-  "STADIUM LOUNGE",
-  "STAGE DOOR/ENCORE",
-  "STALKING HORSE",
-  "STANLEY'S DRIVE INN",
-  "STARBUCKS",
-  "STARBUCKS  COFFEE #746",
-  "STARBUCKS COFFEE",
-  "STARBUCKS COFFEE",
-  "STARBUCKS COFFEE COMPANY",
-  "STARBUCKS CORPORATION",
-  "STARLITE LOUNGE",
-  "STARTERS TAVERN",
-  "STATION NORTH ARTS CAF",
-  "STEELWORKERS HOLDING COMPANY",
-  "STEEPLE CHASE II",
-  "STOCKMARKET BAR",
-  "SUBWAY",
-  "SUBWAY",
-  "subway",
-  "SUBWAY",
-  "SUBWAY",
-  "SUBWAY",
-  "SUBWAY",
-  "SUBWAY  RESTAURANT",
-  "SUBWAY 1301 N. CHARLES ST",
-  "SUGAR HILL TAVERN & PKG.",
-  "SUN FLOWER DELI",
-  "SUN HING CHINESE FOOD CARRYOUT",
-  "SUNLIGHT LOUNGE",
-  "SURN-N-JOE\"S",
-  "SURVIVOR'S SPORTS BAR (606-08)",
-  "SUSHI-SAN RESTAURANT",
-  "SUZIE' SOBA",
-  "SUZIE'S SOBA",
-  "SWEET MELON CAF",
-  "SWEET RETREAT CAF",
-  "SZE-CHUAN RESTAURANT",
-  "T N C. INC. (4106-4108)",
-  "T.D.F. CORPORATION",
-  "TABRIZI'S",
-  "TACO BELL",
-  "TACO BELL & LONG JOHN SILVERS",
-  "TACO FIESTA",
-  "TAKI'S PIZZA CARRYOUT RESTAURANT",
-  "TALAY THAI GRILLE",
-  "TALKING HEAD CLUB",
-  "TAMBER'S",
-  "TAM-TAM",
-  "TAPAS TEATRO CAF",
-  "TASTE",
-  "TASTE INTERNATIONAL RESTAURANT",
-  "TASTE INTERNATIONAL RESTAURANT & LOUNGE/BAR",
-  "TAYLOR'S 5000 TAVERN",
-  "TAYLORS EAST",
-  "TEN-O-SIX",
-  "TEQUILA SUNSET",
-  "THAI  LANDING",
-  "THAI ARROY RESTAURANT",
-  "THAI RESTAURANT",
-  "THE 8 x 10",
-  "THE ADMIRAL'S CUP",
-  "THE AMERICAN LEGION",
-  "THE BALTIMORE SYMPHONY ORCHESTRA",
-  "THE BALTIMORE VETS ASSOC., INC. CHPT #451",
-  "THE BICYCLE",
-  "THE BISTRO",
-  "THE BRASS ELEPHANT",
-  "THE BREWER'S ART",
-  "THE BROOKSHIRE  INNER  HARBOR  SUITE  HOTEL",
-  "THE CANTON LIQUOR HOUSE",
-  "THE CAPITAL GRILLE",
-  "THE CARIBBEAN AMERICAN HEAT",
-  "THE CATS EYE PUB",
-  "The Chameleon Caf‚'",
-  "THE CHEESECAKE  FACTORY",
-  "THE CHINA DOLL",
-  "THE CRAB PIT",
-  "THE DEN LLC",
-  "THE DEPOT",
-  "THE DRINKERY",
-  "THE EDGE BAR & LOUNGE",
-  "THE EDGE BAR & LOUNGE - KITCHEN AREA",
-  "THE ENGINEER'S SOCIETY OF BALTIMORE",
-  "THE EVERGREEN",
-  "THE FAIRMOUNT INN",
-  "THE GARDENVILLE CLUB",
-  "THE GARRETT LOUNGE",
-  "THE GENTLEMAN'S GOLD CLUB",
-  "THE GIN  MILL",
-  "THE GLASS GRILL",
-  "THE GODDESS",
-  "THE GREENE TURTLE",
-  "THE HOLLOW BAR & GRILL",
-  "THE HORSE YOU CAME IN ON",
-  "THE HUBCAP INN",
-  "THE IBIS TAVERN",
-  "THE INNER HARBOR MARINA OF BALTIMORE, INC.",
-  "THE LAUGHING PINT",
-  "THE LIFE OF REILLY IRISH PUB & RESTAURANT",
-  "THE LODGE BAR",
-  "THE LOOP",
-  "THE MORNING EDITION CAFE",
-  "THE NEW 2 O'CLOCK CLUB",
-  "THE NEW BUCKETT'S LOUNGE",
-  "THE NEW CASSELL (KITCHEN AREA)",
-  "THE NEW CASSELL'S",
-  "THE NEW HAVEN LOUNGE",
-  "THE NEW WYMAN PARK RESTAURANT",
-  "THE NORTHEAST TAVERN",
-  "THE OCEANAIRE SEAFOOD ROOM",
-  "THE OPPOSITE SIDEWALK SALOON",
-  "THE PLACE",
-  "THE PREAKNESS LOUNGE",
-  "THE PRIME RIB, INCORPORATED",
-  "THE QUEST",
-  "THE RED STAR BAR & GRILL",
-  "THE ROWAN TREE",
-  "THE SHAMROCK INN",
-  "THE WALTERS ART GALLERY",
-  "THE WATERFRONT HOTEL",
-  "THE WHARF RAT CAMDEN YARDS",
-  "THE WHISPERING WIND, LLC",
-  "THE WINE MARKET (SUITE 135)",
-  "THE YABBA POT",
-  "THIRSTY'S II",
-  "THORNTON'S PUB",
-  "THREE  ACE'S",
-  "THREE KINGS OF EGYPT",
-  "THUMPER'S",
-  "TIBURZI'S CAF LLC",
-  "TIJUANA TACOS & DELI",
-  "TIME CAF",
-  "TIMOTHY DEAN BISTRO",
-  "TIO PEPE RESTAURANTE",
-  "TIR NA NOG BAR & GRILL",
-  "TODD CONNERS",
-  "TOM & OLIVE'S CAFE",
-  "TOMMY'S LOUNGE",
-  "TOM'S BAR",
-  "TONY'S PLACE",
-  "TORAIN'S  HIDE-A-WAY",
-  "TOWER TERRACE RESTAURANT",
-  "TRACEY'S  LOUNGE",
-  "TRAVELERS LOUNGE",
-  "TREMONT GRAND CAF",
-  "TREMONT HOTEL/8 EAST RESTAURANT",
-  "TREMONT PLAZA DELI",
-  "TREMPER INC.",
-  "TRINIDAD ASSOCIATES",
-  "TRIP PLACE",
-  "TROPICAL TASTE RESTAURANT",
-  "TRUFFLES & TEA",
-  "TRUST CHINESE RESTAURANT",
-  "TSUNAMI BALTIMORE",
-  "TUG'S RESTAURANT-TREMONT PLAZA HOTEL",
-  "TURF BAR & LOUNGE",
-  "TURNER'S CAFE",
-  "TURNTABLE LIMITED",
-  "TWINS SUBWAY",
-  "TWO SISTERS GRILLE",
-  "TYSON PLACE, INC.  ( A.K.A 227 W CHASE ST)",
-  "Tyson's Tavern",
-  "TYSON'S TAVERN",
-  "UNCLE LEE'S AT THE HARBOR",
-  "UNCLE WALT'S PLACE",
-  "UNION HOLDING COMPANY",
-  "UNITED AUTO WORKERS HOLDING COMPANY",
-  "UNIVERSAL SOCIAL CLUB",
-  "UPPER DECK",
-  "V I P CITY  (1833-35)",
-  "v.f.w. post 2916",
-  "VACCARO'S ITALIAN PASTRY SHOP, INC.",
-  "VALENTINO'S RESTAURANT,INC",
-  "VELLEGGIA'S ITALIAN SEAFOOD RESTAURANT",
-  "VENICE TAVERN",
-  "VENUS PIZZERIA, INC.",
-  "VERNELL'S ICE PARADISE",
-  "VFW POST #8509  - DONALD CRUMBLE,TRUSTEE",
-  "VIEN BENIDOS LOS CATRACHOS",
-  "VINNY'S  CAFE",
-  "VISION BAR & TAVERN",
-  "VITO'S PIZZA",
-  "VULCAN BLAZERS, INC OF BALTIMORE CITY",
-  "WAGON WHEEL INN",
-  "WALT'S INN",
-  "WAVERLY SOCIAL PLEASURE CLUB",
-  "WEIDOG",
-  "WENDY'S",
-  "WENDY'S",
-  "WENDY'S  RESTAURANT",
-  "WENDY'S #166",
-  "WENDY'S #87",
-  "WENDY'S OLD FASHIONED HAMBURGERS #96",
-  "WENDY'S RESTAURANT",
-  "WERNER'S RESTAURANT",
-  "WESTSIDE INN",
-  "WHISPERS",
-  "WHISTLING OYSTER",
-  "WHITTEN'S  (4502-04)",
-  "WHITTEN'S II",
-  "WILSON'S",
-  "WINDMILL/ MILOS",
-  "WINKS HILLTOP GRILL",
-  "WINSTON LOUNGE",
-  "WOODBERRY KITCHEN",
-  "WOODEN NICKLE LOUNGE",
-  "wozi lounge",
-  "WYLIE  LIQUORS (3101-03)",
-  "XANADU",
-  "XS",
-  "YELLOW BOWL CAF",
-  "YELLOW BOWL RESTAURANT II",
-  "YELLOW DOG TAVERN",
-  "YETI RESTAURANT & CARRYOUT",
-  "YORK CLUB TAVERN",
-  "YOUNG'S PLACE",
-  "ZEEBA LOUNGE",
-  "ZEN WEST ROADSIDE CANTINA",
-  "ZIASCOS",
-  "ZINK'S CAF",
-  "ZISSIMOS BAR",
-  "ZORBAS",
-  "Chungchun Hotdogs",
-  "Buncker21",
-  "Flying Pa-Dak",
-  "Koko BBQ",
-  "Tang 88",
-  "The SmoKING Ribs",
-  "Grams BBQ",
-  "Kaju Tofu",
-  "Pho 45",
-  "Dry Noodle 168",
-  "Hue Oi",
-  "Tam Bien",
-  "Lien Hue",
-  "Bien Hen",
-  "Twinkle Boba",
-  "Sharetea",
-  "7 Leaves Cafe",
-  "Kung Fu Tea",
-  "Gong Cha",
-  "Taste Tea Cafe",
-  "Tasty Noodle House",
-  "Ramen Zuru",
-  "Umaya Ramen",
-  "Tampopo",
-  "Niko Niko",
-  "Kabuki",
-  "Gyu-Kaku",
-  "Dip Shabu Shabu",
-  "Pepper Lunch",
-  "Yuzu Shabu",
-  "Geko Tei",
-  "Nine Seafood",
-  "Claws",
-  "Stinkin Crawfish",
-  "The Cajun Crab",
-  "Louisiana Chicken",
-  "Mezcalero",
-  "The Grasshopper",
-  "The Bamboo Club",
-  "NextDoor",
-  "District Wine",
-  "The Carvery",
-  "555 East",
-  "King Fish House",
-  "The Crab Shack",
-  "Pier 76",
-  "Restauration",
-  "Simmzy",
-  "Crooked Duck",
-  "The Local Spot"
-]
