+++ package.json
@@ -10,7 +10,7 @@
   "scripts": {
     "react": "webpack -d --watch",
     "server": "nodemon server/index.js",
-    "seed-postgres": "node ./database/seeding/pgUsers.js",
+    "seed": "node ./database/seed.js",
     "test": "jest"
   },
   "dependencies": {
@@ -23,14 +23,10 @@
     "axios": "^0.20.0",
     "babel-loader": "^8.1.0",
     "body-parser": "^1.19.0",
-    "csv-write-stream": "^2.0.0",
-    "csv-writer": "^1.6.0",
     "express": "^4.17.1",
     "faker": "^5.1.0",
-    "fs": "0.0.1-security",
     "mongoose": "^5.10.6",
     "nodemon": "^2.0.4",
-    "path": "^0.12.7",
     "postgres": "^1.0.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
